dynamodb-firestore/streaming-replication/lambda-func-firestore/requirements.txt
google-cloud-firestore
boto3
